🎨  moved the authorizationChecker imlementation to helper and updated the DefaultBot Interface
