+++ src/index.ts
@@ -1,21 +1,12 @@
 import express, { Express } from "express";
-import {
-  Action,
-  useExpressServer,
-  UnauthorizedError,
-  BadRequestError,
-  ForbiddenError
-} from "routing-controllers";
+import { Action, useExpressServer } from "routing-controllers";
 import { BotController } from "./controllers";
 import dotenv from "dotenv";
-import { decode } from "jsonwebtoken";
 import * as bodyParser from "body-parser";
 import * as winston from "winston";
 import * as expressWinston from "express-winston";
 import "reflect-metadata";
-import { IJwt } from "./models/jwt";
-import { SecurityMiddleware } from "./middlewares/SecurityMiddleware";
-import configApp from "./configs/app";
+import AuthorizationChecker from "./helpers/AuthorizationChecker";
 dotenv.config();
 
 const app: Express = express();
@@ -32,41 +23,7 @@ app.use(
 );
 
 useExpressServer(app, {
-  authorizationChecker: async (action: Action) => {
-    const authHeader = action.request.headers.authorization;
-
-    if (!authHeader) {
-      throw new ForbiddenError("You are not allowed to do this action");
-    }
-
-    const token = authHeader.split(" ")[1];
-
-    if (!token) {
-      throw new UnauthorizedError("Not authorized to perform this action");
-    }
-
-    const decodedToken = decode(token) as IJwt;
-
-    if (!decodedToken) {
-      throw new BadRequestError("Unable to identify the request");
-    }
-
-    // The following scope will fail if you move to V2 and our
-    // This scope might also be changed in future release of custom endpoint
-    if (
-      decodedToken.scope &&
-      decodedToken.scope !== configApp.CUSTOM_ENDPOINT_APP_SCOPE
-    ) {
-      throw new BadRequestError("Not allowed to do this action");
-    }
-
-    await SecurityMiddleware.validateAuthentication(token, decodedToken);
-
-    // Set the valid auth info to the request chain for access later
-    action.request.auth = decodedToken;
-
-    return true;
-  },
+  authorizationChecker: AuthorizationChecker,
   currentUserChecker: (action: Action) => action.request.auth,
   cors: true,
   routePrefix: "/v1",
